requirements.txt
--extra-index-url https://test.pypi.org/simple/
architect-functions>=0.2.1.dev88
pipdeptree
